restructured the application directory. Now has seperate models, controllers, enums, app, and utils packages
