+++ Hotel/src/models/Guest.java
@@ -1,4 +1,6 @@
-package hotel;
+package models;
+
+import models.User;
 
 public class Guest extends User {
