+++ Hotel/src/models/Rooms.java
@@ -1,4 +1,7 @@
-package hotel;
+package models;
+
+import Enums.BedType;
+import Enums.RoomType;
 
 public class Rooms {
